Add Flash Storage library to Dependencies

+++ WiFi-Motes/AP_UDPtransmitOSCMPU6050.ino/mpu6050_subroutines.ino
@@ -139,12 +139,12 @@ void udp_mpu6050(void)
     bndl.add("/MPU6050/yaw").add((float)(ypr[0] * 180/M_PI));
     bndl.add("/MPU6050/pitch").add((float)(ypr[1] * 180/M_PI));
     bndl.add("/MPU6050/roll").add((float)(ypr[2] * 180/M_PI));
-    bndl.add("/MPU6050/ax").add((uint32_t)ax);
-    bndl.add("/MPU6050/ay").add(ay);
-    bndl.add("/MPU6050/az").add(az);
-    bndl.add("/MPU6050/gx").add(gx);
-    bndl.add("/MPU6050/gy").add(gy);
-    bndl.add("/MPU6050/gz").add(gz);
+    bndl.add("/MPU6050/ax").add((int32_t)ax);
+    bndl.add("/MPU6050/ay").add((int32_t)ay);
+    bndl.add("/MPU6050/az").add((int32_t)az);
+    bndl.add("/MPU6050/gx").add((int32_t)gx);
+    bndl.add("/MPU6050/gy").add((int32_t)gy);
+    bndl.add("/MPU6050/gz").add((int32_t)gz);
 
    // UDP Packet
     Udp.beginPacket("192.168.1.100", 9436);
